source cleanup

+++ src/main/java/com/firemerald/additionalplacements/block/AdditionalWeightedPressurePlateBlock.java
@@ -15,7 +15,7 @@ public class AdditionalWeightedPressurePlateBlock extends AdditionalBasePressure
 	{
 		return new AdditionalWeightedPressurePlateBlock(plate);
 	}
-	
+
 	private AdditionalWeightedPressurePlateBlock(WeightedPressurePlateBlock plate)
 	{
 		super(plate);
@@ -26,7 +26,7 @@ public class AdditionalWeightedPressurePlateBlock extends AdditionalBasePressure
 	{
 		AABB aabb = TOUCH_AABBS[level.getBlockState(pos).getValue(AdditionalBasePressurePlateBlock.PLACING).ordinal() - 1].move(pos);
 		int i = Math.min(level.getEntitiesOfClass(Entity.class, aabb).size(), parentBlock.maxWeight);
-		if (i > 0) return Mth.ceil(15 * (float) Math.min(parentBlock.maxWeight, i) / (float)parentBlock.maxWeight);
+		if (i > 0) return Mth.ceil(15 * (float) Math.min(parentBlock.maxWeight, i) / parentBlock.maxWeight);
 		else return 0;
 	}
 }
